+++ src/duration.cpp
@@ -1,11 +1,15 @@
-#include "clock.h"
-#include "utils.h"
+#include <cpp11/strings.hpp>
+#include <cpp11/list_of.hpp>
+#include <cpp11/integers.hpp>
+#include <cpp11/doubles.hpp>
+#include <cpp11/list.hpp>
+#include <sstream>
+#include <cfloat>
 #include "duration.h"
+#include "utils.h"
 #include "enums.h"
 #include "get.h"
 #include "rcrd.h"
-#include <sstream>
-#include <cfloat>
 
 // -----------------------------------------------------------------------------
