+++ src/gregorian-year-month-day.h
@@ -1,9 +1,12 @@
 #ifndef CLOCK_GREGORIAN_YEAR_MONTH_DAY_H
 #define CLOCK_GREGORIAN_YEAR_MONTH_DAY_H
 
-#include "clock.h"
-#include "integers.h"
+#include <tzdb/date.h>
+#include <cpp11/integers.hpp>
+#include <cpp11/list.hpp>
+#include "r.h"
 #include "enums.h"
+#include "integers.h"
 #include "utils.h"
 #include "stream.h"
 #include "resolve.h"
