+++ src/gregorian-year-month-weekday.h
@@ -1,9 +1,12 @@
 #ifndef CLOCK_GREGORIAN_YEAR_MONTH_WEEKDAY_H
 #define CLOCK_GREGORIAN_YEAR_MONTH_WEEKDAY_H
 
-#include "clock.h"
-#include "integers.h"
+#include <tzdb/date.h>
+#include <cpp11/integers.hpp>
+#include <cpp11/list.hpp>
+#include "r.h"
 #include "enums.h"
+#include "integers.h"
 #include "utils.h"
 #include "stream.h"
 #include "resolve.h"
