+++ src/iso-year-week-day.cpp
@@ -1,3 +1,9 @@
+#include <cpp11/integers.hpp>
+#include <cpp11/logicals.hpp>
+#include <cpp11/list.hpp>
+#include <cpp11/list_of.hpp>
+#include "r.h"
+#include "utils.h"
 #include "iso-year-week-day.h"
 #include "calendar.h"
 #include "duration.h"
