+++ src/parse.h
@@ -1,7 +1,10 @@
 #ifndef CLOCK_PARSE_H
 #define CLOCK_PARSE_H
 
-#include "clock.h"
+#include <tzdb/date.h>
+#include <cpp11/strings.hpp>
+#include <string>
+#include "r.h"
 
 // -----------------------------------------------------------------------------
