+++ src/quarterly-year-quarter-day.h
@@ -1,9 +1,13 @@
 #ifndef CLOCK_QUARTERLY_YEAR_QUARTER_DAY_H
 #define CLOCK_QUARTERLY_YEAR_QUARTER_DAY_H
 
-#include "clock.h"
-#include "integers.h"
+#include <tzdb/date.h>
+#include "quarterly.h"
+#include <cpp11/integers.hpp>
+#include <cpp11/list.hpp>
+#include "r.h"
 #include "enums.h"
+#include "integers.h"
 #include "utils.h"
 #include "stream.h"
 #include "resolve.h"
@@ -85,6 +89,28 @@ inline void check_range<component::microsecond>(const int& value, const char* ar
 template <>
 inline void check_range<component::nanosecond>(const int& value, const char* arg) {
   check_range_nanosecond(value, arg);
+}
+
+enum quarterly::start parse_start(const cpp11::integers& x) {
+  if (x.size() != 1) {
+    clock_abort("`start` must be an integer with length 1.");
+  }
+
+  const int s = x[0];
+
+  if (s == 1) return quarterly::start::january;
+  else if (s == 2) return quarterly::start::february;
+  else if (s == 3) return quarterly::start::march;
+  else if (s == 4) return quarterly::start::april;
+  else if (s == 5) return quarterly::start::may;
+  else if (s == 6) return quarterly::start::june;
+  else if (s == 7) return quarterly::start::july;
+  else if (s == 8) return quarterly::start::august;
+  else if (s == 9) return quarterly::start::september;
+  else if (s == 10) return quarterly::start::october;
+  else if (s == 11) return quarterly::start::november;
+  else if (s == 12) return quarterly::start::december;
+  else clock_abort("'%i' is not a recognized `start` option.", s);
 }
 
 } // namespace detail
