+++ src/r.h
@@ -1,0 +1,18 @@
+#ifndef CLOCK_CLOCK_H
+#define CLOCK_CLOCK_H
+
+// Common utility headers
+#include <R.h>
+#include <Rinternals.h>
+#include <stdint.h>
+#include <stdbool.h>
+
+#define r_null R_NilValue
+#define r_ssize R_xlen_t
+
+#define r_dbl_na NA_REAL
+#define r_int_na NA_INTEGER
+#define r_chr_na NA_STRING
+#define r_lgl_na NA_LOGICAL
+
+#endif
