+++ src/rcrd.cpp
@@ -1,5 +1,5 @@
-#include "rcrd.h"
 #include "utils.h"
+#include "rcrd.h"
 #include "enums.h"
 
 SEXP
