+++ src/rcrd.h
@@ -1,7 +1,7 @@
 #ifndef CLOCK_RCRD_H
 #define CLOCK_RCRD_H
 
-#include "clock.h"
+#include "r.h"
 
 SEXP new_clock_rcrd_from_fields(SEXP fields, SEXP names, SEXP classes);
